+++ 시험/product/src/main/java/com/kh/product/web/form/product/EditForm.java
@@ -11,5 +11,5 @@ public class EditForm {
   private Long pid;       //상품아이디 PID	NUMBER(10,0)
   private String pname;   //상품명 PNAME	VARCHAR2(30 BYTE)
   private Integer count;  //수량 COUNT	NUMBER(10,0)
-  private Integer price;  //가격 PRICE	NUMBER(10,0)
+  private Long price;  //가격 PRICE	NUMBER(10,0)
 }
