+++ 시험/product/src/main/java/com/kh/product/web/form/product/SaveForm.java
@@ -9,5 +9,5 @@ public class SaveForm {
   @NotBlank
   private String pname;   //상품명 PNAME	VARCHAR2(30 BYTE)
   private Integer count;  //수량 COUNT	NUMBER(10,0)
-  private Integer price;  //가격 PRICE	NUMBER(10,0)
+  private Long price;  //가격 PRICE	NUMBER(10,0)
 }
